requirements.txt
fastapi
fastapi-utils
fastapi-cache
pydantic
uvicorn
sqlalchemy
PyJWT
pytz
python-dotenv
python-dateutil
python-multipart
load-dotenv
SQLAlchemy
starlette
bcrypt
passlib
colorama
pytest
httpx
httpcore
certifi
idna
sniffio
requests
urllib3
tzdata
typing_extensions
six
idna
jinja2
